Site: Move Login Here && Add Homepage
The homepage is still work in progress, I lost some time playing with javascript to make the swipe to delete function for the first row, sadly it doesn't even work in -webkit browsers.

More updates will be coming soon to make the page fully functional on the visual side, I still have to give the rows different contents and write the swipe to confirm/edit code.