Angular: Reorganized User & Event Modules
Now they're in separated files on the same level as app-root, this makes it easy for the childs to import them with the angular data binds.

To import the User module inside a component you have to write these lines inside the .ts file:
  import { User } from '../user.module'; // At the start
  @Input() usersDB: User[]; // Inside the class

The same goes for the Event module:
  import { Event } from '../event.module';
  @Input() eventsDB: Event[];

If you want to use it on a new component you have to edit the .ts of the root component binding the data like this inside the component call:
  <app-example [usersDB]="usersDB"</app-example>
  <app-example2 [eventsDB]="eventsDB"</app-example2>

Some things in the code are still hardcoded til we get all the components to work properly.