Angular: Large update, read the description
Routing: removed sessionstatus and implemented angular routing
Database: this is a proper way to get and set the database in all the components
MD5: the events database is saved inside the localstorage, if there's a remote change and the hash is different then it gets updated from remote.
Bootstrap4: updated the classes and fixed all the bugs.
Polyfills: internet explorer 11 is now supported.

There also various fixes and things that don't really deserve a major description.