Angular: Optimize loading time
From 4s to 1s.

The Cuprum font is now stored in the assets, thus is now faster to load because we don't have to make an http request to google.

The database call is now single.

And finally the loading page uses the native ngDoCheck rather than an artificial interval to check if everything's been downloaded.